Change Java version from 1.5 to 1.6 to resolve compilation errors (in Eclipse)
The project did not compile in Eclipse due to the use of @Override annotations in interface method implementations. In Java1.5/JDK5 @Override are not allowed on implemented methods of an interface, only on overridden methods present in a super class. From Java 1.6/JDK6 onwards both usages are supported. See: http://stackoverflow.com/a/2335752/1084488
Strangely this didn't cause problems when compiling the project through mvn, but it did in Eclipse.

Signed-off-by: Matthias Stevens <matthias.stevens@gmail.com>