Added rem(ainder) operation to Operations (& implementing classes) and Arithmetic
The operation computes the remainder of integer division (i.e. a % b). I'm calling it rem (for remainder) instead of the usual mod or modulo because these are not the same in the context of BigInteger. In BigIntegerOperations we use the BigInteger#remainder(BigInteger) and not BigInteger#mod(BigInteger). Hence I'm using the rem(ainder) name everywhere.

Signed-off-by: Matthias Stevens <matthias.stevens@gmail.com>